video_rag_preprocessing/requirements.txt
# Video RAG Preprocessing Pipeline – Dependencies
# Python 3.10+

# Core
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0

# Transformers (CLIP + DINOv2)
transformers>=4.35.0

# Clustering / similarity (used internally by NumPy, available if needed)
scikit-learn>=1.3.0

# Progress
tqdm>=4.65.0

# Image handling
Pillow>=10.0.0

# ANN search
faiss-cpu>=1.7.0

# Signal processing
scipy>=1.10.0

# Config
pyyaml>=6.0

# Validation
pydantic>=2.8.0

# Graph
networkx>=3.1

# Retrieval pipeline
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
